renaming parallelogram to cartesian view implementation
> This is pseudo-amending commit regarding previous commit:d8ae3d4; one important fact to underline, the commit:d8ae3d4 I just mentioned was merged from base to branch:develop with no PR raised on remote, that is uncommon, usually we PR and merge from develop to base, please keep that in mind;
---
-  DEV NOTE In the future, try to consider the combination of grid with isSkewed and the cartesian, it gives a feeling of affine space
