+++ user-settings.js
@@ -14,7 +14,7 @@ export default class {
         Object.assign(context, {
             global: {
                 options: {
-                    scalingValue: 3,
+                    scalingValue: 4,
                     lineWidth: 4,
                     strokeStyle: 'grey',
                 }
